# CUDA: mul_mat_id for mmf for bs <= 64 for f16 and bs <= 32 for f32